215_kamiooya/C1_cursor/browser_automation/requirements.txt
# ブラウザ自動操作（日能研・銀行などログイン後の情報取得）
playwright>=1.40.0
playwright-stealth>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
pdfplumber>=0.10.0
# 東海労金用：自動化検知回避の代替手段（use_undetected_chromedriver: true 時）
undetected-chromedriver>=3.5.0
selenium>=4.0.0
